Update raceregistration.html
Completed the exercise.
The form asks and requires the users first and last name, select a race ("Fun Run 5k" is automatically selected), requires an email in a valid format (containing "@"), a password, and to select ones age. Once everything is filled in according to the guidelines, the button "Register" will take the user to a /registrationValidation page.